requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
appdirs=1.4.4=pypi_0
astor=0.8.1=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blessings=1.7=py36h06a4308_1002
ca-certificates=2022.07.19=h06a4308_0
cachetools=4.2.4=pypi_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
commonmark=0.9.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.24=py36h295c915_0
dataclasses=0.8=pyh4f3eec9_6
datatable=1.0.0=pypi_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.3=py36_0
exgrads=0.1.12=pypi_0
expat=2.4.4=h295c915_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=pypi_0
fsspec=2022.1.0=pypi_0
future=0.18.2=pypi_0
gast=0.5.3=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.18=pypi_0
glib=2.69.1=h4ff587b_1
google-auth=2.6.6=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
gpustat=0.6.0=pyhd3eb1b0_1
grpcio=1.46.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gtasker=1.0.1=pypi_0
gurobipy=9.1.2=pypi_0
h5py=2.10.0=py36h7918eee_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=3.3=pypi_0
idna-ssl=1.1.0=pypi_0
importlib-metadata=4.8.3=pypi_0
importlib-resources=5.4.0=pypi_0
intel-openmp=2022.0.1=h06a4308_3633
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py36_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonrpclib=0.2.1=pypi_0
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.8.1=py36h06a4308_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py36h2531618_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
markdown=3.3.7=pypi_0
matplotlib=3.1.1=py36h5429711_0
mip=1.13.0=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=5.2.0=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
numpy=1.19.5=pypi_0
numpy-base=1.19.2=py36hfa32c7d_0
nvidia-ml=7.352.0=pyhd3eb1b0_0
oauthlib=3.2.0=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=0.25.3=py36he6710b0_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
patsy=0.5.1=py36_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pip=21.2.2=py36h06a4308_0
prettytable=2.5.0=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py3nvml=0.2.7=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.13.0=pypi_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.13=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.10.2=py3.6_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.5.10=pypi_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
pyzmq=22.2.1=py36h295c915_1
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
reckit=0.2.4=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=12.4.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.3.2=py36h7c811a0_0
seaborn=0.11.2=pyhd3eb1b0_0
sentry-sdk=1.5.8=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=59.5.0=pypi_0
shortuuid=1.0.8=pypi_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.38.0=hc218d9a_0
statsmodels=0.12.2=py36h27cfd23_0
tensorboard=1.14.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=1.14.0=pypi_0
tensorflow-estimator=1.14.0=pypi_0
termcolor=1.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
torchmetrics=0.8.2=pypi_0
tornado=6.1=py36h27cfd23_0
tqdm=4.63.0=pyhd3eb1b0_0
traitlets=4.3.3=py36h06a4308_0
typing_extensions=4.1.1=pyh06a4308_0
urllib3=1.26.9=pypi_0
wandb=0.12.11=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
xmltodict=0.12.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
yaspin=2.1.0=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.6.0=pypi_0
zlib=1.2.11=h7f8727e_4
